bug/CORE-1742-broken-sequential-parallels * move stages getnodes into its own file as well
